Backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-channels==4.0.0
channels==4.0.0
channels-redis==4.1.0
djongo==1.3.6
mongoengine==0.27.0
python-dotenv==1.0.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
python-decouple==3.8
drf-spectacular==0.26.5
django-filter==23.5
djangorestframework-jsonapi==6.1.0
python-dateutil==2.8.2
pytz==2023.3
cryptography==41.0.7
pyjwt==2.8.1
